# internal/db/postgres/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql/driver"
	"fmt"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type ConversationType string

const (
	ConversationTypeDm    ConversationType = "dm"
	ConversationTypeGroup ConversationType = "group"
)

func (e *ConversationType) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = ConversationType(s)
	case string:
		*e = ConversationType(s)
	default:
		return fmt.Errorf("unsupported scan type for ConversationType: %T", src)
	}
	return nil
}

type NullConversationType struct {
	ConversationType ConversationType `json:"conversationType"`
	Valid            bool             `json:"valid"` // Valid is true if ConversationType is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullConversationType) Scan(value interface{}) error {
	if value == nil {
		ns.ConversationType, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.ConversationType.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullConversationType) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.ConversationType), nil
}

type Conversation struct {
	ID        uuid.UUID          `db:"id" json:"id"`
	Type      ConversationType   `db:"type" json:"type"`
	Title     pgtype.Text        `db:"title" json:"title"`
	Slug      pgtype.Text        `db:"slug" json:"slug"`
	CreatedAt pgtype.Timestamptz `db:"created_at" json:"createdAt"`
}

type ConversationMember struct {
	ID             uuid.UUID          `db:"id" json:"id"`
	ConversationID uuid.UUID          `db:"conversation_id" json:"conversationId"`
	UserID         uuid.UUID          `db:"user_id" json:"userId"`
	JoinedAt       pgtype.Timestamptz `db:"joined_at" json:"joinedAt"`
	LastReadAt     pgtype.Timestamptz `db:"last_read_at" json:"lastReadAt"`
}

type DmPair struct {
	UserLow        uuid.UUID `db:"user_low" json:"userLow"`
	UserHigh       uuid.UUID `db:"user_high" json:"userHigh"`
	ConversationID uuid.UUID `db:"conversation_id" json:"conversationId"`
}

type Message struct {
	ID             uuid.UUID          `db:"id" json:"id"`
	ConversationID uuid.UUID          `db:"conversation_id" json:"conversationId"`
	SenderID       uuid.UUID          `db:"sender_id" json:"senderId"`
	Content        string             `db:"content" json:"content"`
	CreatedAt      pgtype.Timestamptz `db:"created_at" json:"createdAt"`
}

type PhoneOtpCode struct {
	Phone     string             `db:"phone" json:"phone"`
	CodeHash  string             `db:"code_hash" json:"codeHash"`
	ExpiresAt pgtype.Timestamptz `db:"expires_at" json:"expiresAt"`
	Attempts  int32              `db:"attempts" json:"attempts"`
	SentAt    pgtype.Timestamptz `db:"sent_at" json:"sentAt"`
	CreatedAt pgtype.Timestamptz `db:"created_at" json:"createdAt"`
}

type User struct {
	ID                 uuid.UUID          `db:"id" json:"id"`
	Phone              string             `db:"phone" json:"phone"`
	Username           pgtype.Text        `db:"username" json:"username"`
	PasswordHash       pgtype.Text        `db:"password_hash" json:"passwordHash"`
	AvatarUrl          pgtype.Text        `db:"avatar_url" json:"avatarUrl"`
	CreatedAt          pgtype.Timestamptz `db:"created_at" json:"createdAt"`
	UsernameSearchable bool               `db:"username_searchable" json:"usernameSearchable"`
	DmPolicy           string             `db:"dm_policy" json:"dmPolicy"`
}